common: Move FFT logic to separate fft module
