+++ Makefile
@@ -10,7 +10,7 @@ GUI_BIN = slowrx
 
 TARGETS = $(GUI_BIN)
 
-SOURCES = common.c gui_events.c modespec.c gui.c video.c vis.c sync.c pcm.c fsk.c slowrx.c
+SOURCES = common.c gui_events.c fft.c modespec.c gui.c video.c vis.c sync.c pcm.c fsk.c slowrx.c
 OBJECTS = $(patsubst %.c,%.o,$(SOURCES))
 DEPENDS = $(patsubst %.c,%.d,$(SOURCES))
