+++ fft.c
@@ -1,0 +1,5 @@
+#include <fftw3.h>
+
+#include "fft.h"
+
+FFTStuff     fft;
