+++ fsk.c
@@ -6,6 +6,7 @@
 #include <fftw3.h>
 
 #include "common.h"
+#include "fft.h"
 #include "fsk.h"
 #include "pcm.h"
